Changed code to use the Permission object, it's a bit cleaner now
